Some token pool changes
