+++ scripts/utils/addresses.js
@@ -66,10 +66,13 @@ const sepolia = {
   endemicErc721: '0xa4383Eb3f2Ca59FA51871A19CF3D753235f9041D',
   paymentManagerProxy: '0xe5C828a0AB005E1f2403600C070357c413D313Ba', //implementation: 0x5749624215bCab94b1D09d5fdc81B8Ba2a33AA20
   endemicExchangeProxy: '0x00BfA98843a6d855d44Bff622C3cFA130F720CaD', //implementation: 0x1514B11348ADA62B45fb4ec235E9980662a9Ef21
+  endemicToken: '0xC10B2Bfb5392bEB090570d53A488fC41efFb3491',
 
   artOrders: '0x51f37aCE1c7a05B4c2a90a23E05058E1a6202D7C', //implementation: 0x2F6FB5e4a88CDCFB2F0dDa89FB58671812ca2754
   artOrderFactory: '0x8E3C22767df3164cD80b1802b717f1f41193f6aE', //implementation: 0x5B073C9A381146593f943b6E821F5eE220847783
   artOrderCollection: '0xC1E23C1b1744eD7f1d6EA8b374CC0085938106cC',
+
+  endemicTokenLockingPool: '0x6571bCc2412CD2255b3e74b1a481C5991c00ecBd',
 };
 
 const mumbai = {
